+++ REngine.RPI/Features/PostProcess/GrayscalePostProcess.cs
@@ -1,0 +1,20 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using REngine.Core;
+using REngine.Core.IO;
+
+namespace REngine.RPI.Features.PostProcess
+{
+    public class GrayscalePostProcess : PostProcessFeature
+    {
+        protected override ShaderStream OnGetShaderCode()
+        {
+            return new FileShaderStream(
+                Path.Join(EngineSettings.AssetsShadersPath, "grayscale_ps.hlsl")
+            );
+        }
+    }
+}
