+++ REngine.RPI/RenderGraph/Module.cs
@@ -3,11 +3,12 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using REngine.RPI.RenderGraph.PostProcess;
 
 namespace REngine.RPI.RenderGraph
 {
 #if RENGINE_RENDERGRAPH
-	public static class NodeGraphsModule
+    public static class NodeGraphsModule
 	{
 		public static void Setup(RenderGraphRegistry registry)
 		{
@@ -21,7 +22,8 @@ namespace REngine.RPI.RenderGraph
 				.Register<DepthBufferNode>()
 				.Register<ClearNode>()
 				.Register<CopyRenderTargetNode>()
-				.Register<GrayScalePostProcessNode>();
+				.Register<GrayScalePostProcessNode>()
+				.Register<InvertPostProcessNode>();
 		}
 	}
 #endif
